Append "exec" to java command
PaperMC does not receive the SIGTERM signal from docker when attempting to stop the container. This causes docker to kill the container after 10 seconds and the server is not shut down gracefully. Adding "exec" causes the java command to become PID 1 instead of papermc.sh, then the SIGTERM signal is sent directly to java